refactor: разделить хранилище ревью на два: для ревью и для лайков

+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -13,7 +13,7 @@ public interface ReviewService {
 
     Review getByID(long reviewID);
 
-    List<Review> getAllByFilmID(Integer filmID, int count);
+    List<Review> getAllByFilmID(Long filmID, int count);
 
     void addLike(long reviewID, long userID);
